requirements.txt
aiohttp 
aiosignal 
altair 
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio
asttokens 
async-timeout 
attrs 
backcall 
backoff==2.2.1
bcrypt==4.0.1
beautifulsoup4==4.12.2
blinker 
blis==0.7.11
Bottleneck 
Brotli 
cachetools 
catalogue==2.0.10
certifi
cffi 
chardet==5.2.0
charset-normalizer 
chroma-hnswlib==0.7.3
chromadb==0.4.14
click 
cloudpathlib==0.16.0
colorama 
coloredlogs==15.0.1
comm 
confection==0.1.3
contourpy==1.1.1
cryptography 
cycler==0.12.1
cymem==2.0.8
Cython==3.0.3
dataclasses-json 
debugpy 
decorator 
EbookLib==0.18
effdet==0.4.1
emoji==2.8.0
et-xmlfile==1.1.0
exceptiongroup 
executing 
faiss-cpu==1.7.4
fastapi==0.103.2
filelock==3.12.4
filetype==1.2.0
flatbuffers==23.5.26
fonttools==4.43.1
frozenlist 
fsspec==2023.9.2
gitdb 
GitPython 
google-api-core 
google-auth 
googleapis-common-protos 
googlesearch-python==1.2.3
greenlet 
grpcio==1.59.0
h11==0.14.0
httptools==0.6.0
huggingface-hub==0.17.3
humanfriendly==10.0
idna 
importlib-metadata 
importlib-resources==6.1.0
InstructorEmbedding==1.0.1
iopath==0.1.10
ipykernel 
ipython 
ipywidgets 
jedi 
jellyfish==1.0.1
Jinja2 
joblib==1.3.2
jsonpatch 
jsonpointer 
jsonschema 
jupyter_client 
jupyter_core 
jupyterlab-widgets 
kiwisolver==1.4.5
langchain 
langcodes==3.3.0
langdetect==1.0.9
langsmith 
layoutparser==0.3.4
lxml==4.9.3
Markdown==3.5
markdown-it-py 
MarkupSafe 
marshmallow 
marshmallow-enum 
matplotlib==3.8.0
matplotlib-inline 
mdurl 
mkl-service==2.4.0
monotonic==1.6
mpmath==1.3.0
msg-parser==1.2.0
multidict 
murmurhash==1.0.10
mypy-extensions 
nest-asyncio 
networkx==3.1
nltk==3.8.1
NRCLex==3.0.0
numexpr 
numpy==1.26.1
olefile==0.46
omegaconf==2.3.0
onnx==1.14.1
onnxruntime==1.15.1
openai==0.28.1
openapi-schema-pydantic 
opencv-python==4.8.1.78
openpyxl==3.1.2
overrides==7.4.0
packaging 
pandas 
parso 
pdf2image==1.16.3
pdfminer.six==20221105
pdfplumber==0.10.2
pickleshare 
Pillow==10.1.0
platformdirs 
portalocker==2.8.2
posthog==3.0.2
preshed==3.0.9
prompt-toolkit 
protobuf
psutil 
pulsar-client==3.3.0
pure-eval 
pyarrow==11.0.0
pyasn1 
pyasn1-modules 
pycocotools==2.0.7
pycparser 
pydantic 
pydantic_core==2.10.1
pydeck==0.8.1b0
Pygments 
Pympler 
pyOpenSSL 
pypandoc==1.11
pyparsing==3.1.1
PyPDF2==3.0.1
pypdfium2==4.21.0
PyPika==0.48.9
pyreadline3==3.4.1
pyrsistent 
PySocks 
pytesseract==0.3.10
python-dateutil 
python-decouple==3.8
python-docx==1.0.1
python-dotenv 
python-iso639==2023.6.15
python-magic==0.4.27
python-multipart==0.0.6
python-pptx==0.6.21
pytrends==4.9.2
pytz 
pyu2f 
PyYAML==6.0.1
pyzmq 
rapidfuzz==3.4.0
regex 
requests 
rich 
rsa 
safetensors==0.4.0
scikit-learn==1.3.1
scipy==1.11.3
segtok==1.5.11
sentence-transformers==2.2.2
sentencepiece==0.1.99
six 
smart-open==6.4.0
smmap 
sniffio 
soupsieve==2.5
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy 
srsly==2.4.8
stack-data 
starlette==0.27.0
streamlit==1.27.2
stringcase==1.2.0
sympy==1.12
tabulate==0.9.0
tenacity 
textblob==0.17.1
thinc==8.2.1
threadpoolctl==3.2.0
tiktoken 
timm==0.9.7
tokenizers==0.14.1
toml 
toolz 
torch==2.1.0
torchvision==0.16.0
tornado 
tqdm==4.66.1
traitlets 
transformers==4.34.0
typer==0.9.0
typing-inspect 
typing_extensions 
tzdata 
tzlocal 
unstructured==0.10.23
unstructured-inference==0.7.5
unstructured.pytesseract==0.3.12
urllib3 
uvicorn==0.23.2
vaderSentiment==3.3.2
validators 
wasabi==1.1.2
watchdog 
watchfiles==0.21.0
wcwidth 
weasel==0.3.3
websockets==11.0.3
widgetsnbextension 
win-inet-pton 
xlrd==2.0.1
XlsxWriter==3.1.8
yake==0.4.8
yarl 
zipp 

